wip initial fixes on types in set reg and get reg, cp hll setReg to a.c

+++ src/hll.h
@@ -4,7 +4,7 @@ static inline uint8_t clz(uint64_t x);
 static inline double sigma(double x);
 static inline double tau(double x);
 
-static inline void setReg(uint64_t m, uint8_t n, char *regs);
+static inline void setReg(uint64_t m, uint64_t n, char *regs);
 static inline uint64_t getReg(uint64_t m, char * regs);
 
 void printByte(char a);
